feat(cloudquery): Collect data for AWS lambdas
